+++ transport/src/main/java/com/cooperative/transport/repositories/AnnulationRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cooperative.transport.entities.Annulations;
 
-public interface AnnulationRepository extends JpaRepository<Annulations, Long> {
+public interface AnnulationRepository extends JpaRepository<Annulations, Integer> {
 }
