+++ transport/src/main/java/com/cooperative/transport/repositories/ReservationFilleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.cooperative.transport.entities.ReservationsFille;
 import com.cooperative.transport.entities.ReservationsMere;
 
-public interface ReservationFilleRepository extends JpaRepository<ReservationsFille, Long> {
+public interface ReservationFilleRepository extends JpaRepository<ReservationsFille, Integer> {
     List<ReservationsFille> findByReservationMere(ReservationsMere reservationMere);
 }
