+++ transport/src/main/java/com/cooperative/transport/repositories/StatutPaiementRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cooperative.transport.entities.StatutPaiement;
 
-public interface StatutPaiementRepository extends JpaRepository<StatutPaiement, Long> {
+public interface StatutPaiementRepository extends JpaRepository<StatutPaiement, Integer> {
 
     public Optional<StatutPaiement> findByLibelle(String libelle);
 }
